Big refactor of large modules
