+++ examples/beaker/configs/large_model.yaml
