+++ examples/beaker/configs/mixed_gpu.yaml
