+++ examples/beaker/configs/model_comparison.yaml
