+++ src/olmo_eval/cli/beaker/launch.py
@@ -1,4 +1,14 @@
-"""Launch command for Beaker jobs."""
+"""Launch command for Beaker jobs.
+
+This module provides the main 'beaker launch' command for submitting evaluation jobs.
+Configuration loading, validation, and job assembly are delegated to:
+- config_loader.py: LaunchConfigLoader for loading/merging config
+- task_validator.py: TaskValidator for task validation and priority grouping
+- credentials.py: CredentialManager for AWS/GCS credential handling
+- model_grouper.py: ModelGrouper for grouping models by runtime signature
+- experiment_builder.py: ExperimentPlanBuilder for building experiment plans
+- job_assembler.py: JobConfigAssembler for assembling BeakerJobConfig
+"""
 
 import click
 from rich.panel import Panel
@@ -13,7 +23,7 @@ from olmo_eval.cli.utils import (
     console,
     parse_model_spec,
 )
-from olmo_eval.core.constants.infrastructure import BEAKER_RESULT_DIR, DEFAULT_MAX_GPUS_PER_NODE
+from olmo_eval.core.constants.infrastructure import BEAKER_RESULT_DIR
 
 
 @click.command()
@@ -184,49 +194,11 @@ def launch(
     Models with compatible runtime configurations (GPUs, provider, etc.) are grouped together.
     Use --config/-f to load settings from a YAML file; CLI arguments override config values.
     Use --group/-g to organize experiments into a Beaker group for result aggregation.
-
-    Inference provider packages (vllm, transformers, etc.) are auto-detected from each
-    model's configuration and installed at job startup. Specify per-model providers in
-    the config file using the 'provider' field on each model.
-
-    Examples:
-
-        olmo-eval beaker launch -n "eval-llama3" -m llama3.1-8b -t mmlu
-
-        olmo-eval beaker launch -n "eval-suite" -m llama3.1-8b -t mmlu -t gsm8k -t arc
-
-        olmo-eval beaker launch -n "eval-70b" -m llama3.1-70b -t mmlu --cluster h100 --gpus 4
-
-        # Multiple models (grouped if compatible runtime)
-        olmo-eval beaker launch -n "eval-compare" -m llama3.1-8b -m olmo-2-7b -t mmlu -t gsm8k
-
-        # Per-task priorities (creates separate experiments per priority level)
-        olmo-eval beaker launch -n "eval-mixed" -m llama3.1-8b -t "mmlu@high" -t "gsm8k@normal"
-
-        # From YAML config file (with per-model provider configuration)
-        olmo-eval beaker launch -f eval_config.yaml
-
-        # Config file with CLI overrides
-        olmo-eval beaker launch -f eval_config.yaml --gpus 4 --priority high
-
-        # With grouping for result aggregation
-        olmo-eval beaker launch -n "benchmark" --group "benchmark-2024" \\
-            -m llama3.1-8b -t mmlu -t gsm8k
     """
-    import json as json_module
+    from datetime import datetime
 
     try:
-        from olmo_eval.launch import (
-            BeakerEnvSecret,
-            BeakerJobConfig,
-            BeakerLauncher,
-            EvalConfig,
-            ModelConfig,
-            calculate_experiment_splits,
-            get_model_short_name,
-            parse_model_config,
-            validate_priority_configuration,
-        )
+        from olmo_eval.launch import BeakerLauncher, EvalConfig
     except ImportError:
         console.print(
             "[red]beaker-py is not installed.[/red]\n"
@@ -234,299 +206,217 @@ def launch(
         )
         raise SystemExit(1) from None
 
-    # Track which CLI args were explicitly set (vs using defaults)
-    cli_cluster = cluster
-    cli_gpus = gpus
-    cli_parallelism = parallelism
-    cli_priority = priority
-    cli_preemptible = preemptible
-    cli_timeout = timeout
+    from olmo_eval.cli.beaker.config_loader import LaunchConfigLoader
+    from olmo_eval.cli.beaker.credentials import CredentialManager
+    from olmo_eval.cli.beaker.experiment_builder import ExperimentPlanBuilder
+    from olmo_eval.cli.beaker.job_assembler import JobConfigAssembler
+    from olmo_eval.cli.beaker.model_grouper import ModelGrouper
+    from olmo_eval.cli.beaker.task_validator import TaskValidator
+    from olmo_eval.core.constants.infrastructure import BEAKER_DEFAULT_IMAGE
 
-    # Load config from file if provided
-    cfg: EvalConfig | None = None
-    model_configs: list[ModelConfig] = []
+    # Build CLI args dict
+    cli_args = {
+        "name": name,
+        "model": model,
+        "task": task,
+        "cluster": cluster,
+        "gpus": gpus,
+        "parallelism": parallelism,
+        "max_gpus_per_node": max_gpus_per_node,
+        "priority": priority,
+        "preemptible": preemptible,
+        "timeout": timeout,
+        "retries": retries,
+        "workspace": workspace,
+        "budget": budget,
+        "image": image,
+        "group": group,
+        "use_async": use_async,
+        "use_async_stream": use_async_stream,
+        "num_workers": num_workers,
+        "gpus_per_worker": gpus_per_worker,
+        "s3_bucket": s3_bucket,
+        "s3_prefix": s3_prefix,
+        "s3_endpoint_url": s3_endpoint_url,
+        "s3_region": s3_region,
+        "store": store,
+        "debug_requests": debug_requests,
+        "debug_provider": debug_provider,
+        "save_predictions": save_predictions,
+        "save_requests": save_requests,
+    }
 
+    # Load configuration
+    config_loader = LaunchConfigLoader(config, cli_args)
+    launch_config = config_loader.load()
+
+    # Load EvalConfig for resource lookups
+    eval_config: EvalConfig | None = None
     if config:
-        try:
-            cfg = EvalConfig.from_yaml(config)
-        except FileNotFoundError as e:
-            console.print(f"[red]Error:[/red] {e}")
-            raise SystemExit(1) from None
-        except Exception as e:
-            console.print(f"[red]Config error:[/red] {e}")
-            raise SystemExit(1) from None
+        import contextlib
 
-        # Use config values as defaults, CLI args override
-        name = name or cfg.name
-        task = task if task else tuple(cfg.tasks)
-        retries = retries if retries is not None else cfg.retries
-        workspace = workspace or cfg.workspace
-        budget = budget or cfg.budget
+        with contextlib.suppress(Exception):
+            eval_config = EvalConfig.from_yaml(config)
 
-        # Get model configs from file (with per-model resource overrides)
-        # Also store the original specs (with ::overrides) for command building
-        if not model:
-            model_configs = cfg.get_model_configs()
-            # For config file models, use name_or_path as spec (no inline overrides)
-            model_specs: list[str] = [m.name_or_path for m in model_configs]
-        else:
-            # CLI models override config file models
-            model_specs = list(model)  # Original specs with ::overrides
-            model_configs = [parse_model_config(m) for m in model]
+    # Validate tasks and group by priority
+    task_validator = TaskValidator(
+        launch_config.task_specs,
+        cli_priority=priority,
+        default_priority=launch_config.priority,
+    )
+    tasks_by_priority, valid_tasks, agent_task_specs = task_validator.validate_and_group()
 
-        # Set defaults from config (will be overridden by per-model or CLI)
-        cluster = cluster if cluster is not None else cfg.cluster
-        gpus = gpus if gpus is not None else cfg.gpus
-        parallelism = parallelism if parallelism is not None else cfg.parallelism
-        if max_gpus_per_node is None:
-            max_gpus_per_node = cfg.max_gpus_per_node
-        priority = priority if priority is not None else cfg.priority
-        preemptible = preemptible if preemptible is not None else cfg.preemptible
-        timeout = timeout if timeout is not None else cfg.timeout
-        use_async = use_async or cfg.use_async
-        num_workers = num_workers if num_workers is not None else cfg.num_workers
-        gpus_per_worker = gpus_per_worker if gpus_per_worker != 1 else cfg.gpus_per_worker
-    else:
-        # No config file - use CLI models
-        model_specs = list(model) if model else []  # Original specs with ::overrides
-        model_configs = [parse_model_config(m) for m in model] if model else []
+    # Create launcher
+    launcher = BeakerLauncher(workspace=launch_config.workspace)
 
-    # Apply defaults for values not set by config or CLI
-    gpus = gpus if gpus is not None else 1
-    parallelism = parallelism if parallelism is not None else 1
-    if max_gpus_per_node is None:
-        max_gpus_per_node = DEFAULT_MAX_GPUS_PER_NODE
-    priority = priority or "normal"
-    preemptible = preemptible if preemptible is not None else True
-    timeout = timeout or "24h"
+    # Set up credentials
+    cred_manager = CredentialManager(
+        launch_config.model_configs,
+        launch_config.store,
+        aws_credentials,
+        gcs_credentials,
+    )
+    inject_aws, inject_gcs = cred_manager.detect_and_setup(launcher)
 
-    # Validate required fields
-    if not name:
-        console.print("[red]Error:[/red] --name/-n is required (or set 'name' in config)")
-        raise SystemExit(1) from None
-    if not model_configs:
-        console.print("[red]Error:[/red] --model/-m is required (or set 'models' in config)")
-        raise SystemExit(1) from None
-    if not task:
-        console.print("[red]Error:[/red] --task/-t is required (or set 'tasks' in config)")
-        raise SystemExit(1) from None
-    if not cluster:
-        console.print("[red]Error:[/red] --cluster/-c is required (or set 'cluster' in config)")
-        raise SystemExit(1) from None
-    if not workspace:
-        console.print("[red]Error:[/red] --workspace/-w is required (or set 'workspace' in config)")
-        raise SystemExit(1) from None
-    if not budget:
-        console.print("[red]Error:[/red] --budget/-B is required (or set 'budget' in config)")
-        raise SystemExit(1) from None
-
-    # Validate S3 options - both bucket and prefix required if either is set
-    if s3_bucket and not s3_prefix:
-        console.print("[red]Error:[/red] --s3-prefix is required when --s3-bucket is set")
-        raise SystemExit(1) from None
-    if s3_prefix and not s3_bucket:
-        console.print("[red]Error:[/red] --s3-bucket is required when --s3-prefix is set")
-        raise SystemExit(1) from None
-
-    # --store requires S3 configuration for storing result artifacts
-    if store and (not s3_bucket or not s3_prefix):
-        console.print(
-            "[red]Error:[/red] --s3-bucket and --s3-prefix are required when --store is enabled"
-        )
-        raise SystemExit(1) from None
-
-    # Keep suites unexpanded - let the runner expand them so it knows suite names for aggregation
-    from olmo_eval.core.configs import expand_tasks, validate_tasks
-
-    original_task_specs = list(task)  # Preserve original suite/task names
-
-    # Group by priority WITHOUT expanding first - this keeps suites as single units
-    try:
-        tasks_by_priority = validate_priority_configuration(
-            tasks=original_task_specs,
-            cli_priority=cli_priority,
-            default_priority=priority,
-        )
-    except ValueError as e:
-        console.print(f"[red]Error:[/red] {e}")
-        raise SystemExit(1) from None
-
-    # Get all specs (without @priority suffix, but with ::overrides)
-    all_task_specs = [t for tasks in tasks_by_priority.values() for t in tasks]
-
-    # Expand for validation only - ensure all tasks/suites exist
-    expanded_for_validation = expand_tasks(all_task_specs)
-    valid_tasks, invalid_tasks = validate_tasks(expanded_for_validation)
-
-    # Detect agent tasks vs simple tasks
-    from olmo_eval.evals.tasks import AgentTask
-    from olmo_eval.evals.tasks import get_task as get_task_for_classification
-
-    agent_task_specs: set[str] = set()
-    for task_spec in expanded_for_validation:
-        try:
-            task_instance = get_task_for_classification(task_spec)
-            if isinstance(task_instance, AgentTask):
-                # Strip overrides to match how is_agent_spec() checks
-                base_spec = task_spec.split("::", 1)[0]
-                agent_task_specs.add(base_spec)
-        except Exception:
-            pass  # Skip tasks that can't be loaded - they'll fail validation
-
-    # Classify original specs (with priority suffixes) as agent or simple
-    def is_agent_spec(spec: str) -> bool:
-        """Check if a task spec (potentially with priority suffix) is an agent task."""
-        from olmo_eval.evals.tasks import get_base_task_name
-
-        base_spec = get_base_task_name(spec)
-        expanded = expand_tasks([base_spec])
-        return all(t in agent_task_specs for t in expanded)
-
-    # Track expanded task counts per priority for display purposes
-    expanded_counts_by_priority: dict[str, int] = {}
-    for priority_level, specs in tasks_by_priority.items():
-        expanded_counts_by_priority[priority_level] = len(expand_tasks(specs))
-
-    if invalid_tasks:
-        console.print("[red]Error:[/red] The following tasks/suites do not exist:")
-        for inv in invalid_tasks:
-            console.print(f"  - {inv}")
-        console.print("\nUse 'olmo-eval tasks' to see available tasks.")
-        console.print("Use 'olmo-eval suites' to see available suites.")
-        raise SystemExit(1) from None
-
-    launcher = BeakerLauncher(workspace=workspace)
-    multiple_models = len(model_configs) > 1
-    multiple_priorities = len(tasks_by_priority) > 1
-
-    # Auto-detect when AWS credentials are needed
-    from olmo_eval.launch.beaker.aws import get_local_aws_credentials, is_s3_path
-
-    s3_models = [m.name_or_path for m in model_configs if is_s3_path(m.name_or_path)]
-    inject_aws_credentials = aws_credentials
-    if inject_aws_credentials is None:
-        inject_aws_credentials = bool(s3_models) or store
-
-    # Auto-detect GCS model paths for GCS credential injection
-    from olmo_eval.launch.beaker.gcs import get_local_gcs_credentials, is_gcs_path
-
-    gcs_models = [m.name_or_path for m in model_configs if is_gcs_path(m.name_or_path)]
-    inject_gcs_credentials = gcs_credentials
-    if inject_gcs_credentials is None:
-        inject_gcs_credentials = bool(gcs_models)
-
-    if inject_gcs_credentials:
-        local_gcs_creds = get_local_gcs_credentials()
-        beaker_user = launcher.beaker.user_name
-
-        gcs_table = Table(show_header=False, box=None, expand=True)
-        gcs_table.add_column("Key", style="blue")
-        gcs_table.add_column("Value")
-
-        if local_gcs_creds:
-            gcs_table.add_row("Credentials", "[green]found[/green] (service account)")
-            if local_gcs_creds.client_email:
-                gcs_table.add_row("Service account", local_gcs_creds.client_email)
-            if local_gcs_creds.project_id:
-                gcs_table.add_row("Project", local_gcs_creds.project_id)
-            gcs_table.add_row("Beaker user", beaker_user)
-            gcs_table.add_row("Beaker secret", f"{beaker_user}_GOOGLE_CREDENTIALS")
-        else:
-            gcs_table.add_row(
-                "Credentials",
-                "[yellow]not found[/yellow] - job may fail if GCS access is required",
-            )
-
-        console.print()
-        console.print(
-            Panel(gcs_table, title="[bold]GCS Access Configuration[/bold]", border_style="magenta")
-        )
-        console.print()
-
-    # Get workspace object for beaker API calls that require it
-    workspace_obj = launcher.beaker.workspace.get(workspace) if workspace else None
+    # Update config with credential settings
+    launch_config.inject_aws_credentials = inject_aws
+    launch_config.inject_gcs_credentials = inject_gcs
 
     if dry_run:
         console.print("[yellow]Dry run mode - not submitting[/yellow]")
 
-    # Build list of groups from CLI and config
-    from datetime import datetime
-
-    effective_groups: list[str] = list(group)
-    if cfg is not None and cfg.groups:
-        for g in cfg.groups:
-            if g not in effective_groups:
-                effective_groups.append(g)
-
-    # Auto-generate a group if none specified
+    # Auto-generate group if needed
+    effective_groups = list(launch_config.groups)
     if not effective_groups:
-        effective_groups = [f"{name}-{datetime.now().strftime('%Y%m%d-%H%M%S')}"]
+        effective_groups = [f"{launch_config.name}-{datetime.now().strftime('%Y%m%d-%H%M%S')}"]
 
-    # Display storage configuration if enabled
-    if store or (s3_bucket and s3_prefix) or inject_aws_credentials:
-        storage_lines = []
+    # Display storage info
+    cred_manager.display_storage_info(
+        launcher,
+        launch_config.s3_bucket,
+        launch_config.s3_prefix,
+        launch_config.s3_region,
+        launch_config.s3_endpoint_url,
+        effective_groups,
+        inject_aws,
+    )
 
-        # S3 Access credentials
-        if inject_aws_credentials:
-            local_creds = get_local_aws_credentials()
-            beaker_user = launcher.beaker.user_name
-            storage_lines.append("[bold]S3 Access:[/bold]")
-            if local_creds:
-                cred_type = "temporary" if local_creds.session_token else "long-term"
-                storage_lines.append(f"  Credentials: [green]found[/green] ({cred_type})")
-                storage_lines.append(
-                    f"  Beaker secrets: {beaker_user}_AWS_ACCESS_KEY_ID, "
-                    f"{beaker_user}_AWS_SECRET_ACCESS_KEY"
-                )
-            else:
-                storage_lines.append(
-                    "  Credentials: [yellow]not found[/yellow] - "
-                    "job may fail if S3 access is required"
-                )
+    # Determine effective image
+    if image:
+        effective_image = image
+    elif eval_config and eval_config.beaker_image:
+        effective_image = eval_config.beaker_image
+    else:
+        effective_image = BEAKER_DEFAULT_IMAGE
 
-        # S3 storage settings
-        if s3_bucket and s3_prefix:
-            storage_lines.append("[bold]S3 Storage:[/bold]")
-            storage_lines.append(f"  Bucket: {s3_bucket}")
-            storage_lines.append(f"  Prefix: {s3_prefix}")
-            storage_lines.append(f"  Region: {s3_region}")
-            if s3_endpoint_url:
-                storage_lines.append(f"  Endpoint: {s3_endpoint_url}")
-            if effective_groups:
-                storage_lines.append(f"  Group: {effective_groups[0]}")
+    # Handle group creation
+    _handle_group_creation(launcher, effective_groups, dry_run)
 
-        # PostgreSQL storage
-        if store:
-            storage_lines.append("[bold]PostgreSQL:[/bold]")
-            storage_lines.append(
-                "  Credentials from Beaker secrets: olmo_eval_PGHOST, olmo_eval_PGPORT,"
-            )
-            storage_lines.append("    olmo_eval_PGDATABASE, olmo_eval_PGUSER, olmo_eval_PGPASSWORD")
+    # Group models and build experiment plan
+    model_grouper = ModelGrouper(launch_config, eval_config)
+    experiment_builder = ExperimentPlanBuilder(
+        launch_config, model_grouper, tasks_by_priority, agent_task_specs
+    )
+    experiment_plan, split_models = experiment_builder.build()
 
+    # Get task summaries
+    task_summaries = _get_task_summaries(valid_tasks)
+    task_summary_by_spec = {
+        (ts.spec if ts.spec and ts.spec != ts.config.name else ts.config.name): ts
+        for ts in task_summaries
+    }
+
+    # Collect required secrets
+    all_required_secrets: set[str] = set()
+    for ts in task_summaries:
+        if hasattr(ts.config, "required_secrets") and ts.config.required_secrets:
+            all_required_secrets.update(ts.config.required_secrets)
+
+    # Ensure secrets
+    common_secrets, store_secrets, task_secrets = _ensure_secrets(
+        launcher, dry_run, launch_config, all_required_secrets
+    )
+
+    # Print summary header
+    total_experiments = len(experiment_plan)
+    total_expanded_tasks = len(valid_tasks) * len(launch_config.model_configs)
+    console.print()
+    console.print(
+        f"[bold]Launching {total_experiments} experiment(s) "
+        f"with {total_expanded_tasks} task(s)[/bold]"
+    )
+    if split_models:
+        console.print(
+            "[dim]  Tasks distributed across multiple experiments due to GPU constraints[/dim]"
+        )
+
+    # Show experiment matrix if multiple experiments
+    if total_experiments > 1:
+        _print_experiment_matrix(experiment_plan, use_async_stream, use_async)
+
+    console.print()
+
+    # Build job configs and summaries
+    job_assembler = JobConfigAssembler(
+        launch_config,
+        eval_config,
+        effective_image,
+        effective_groups,
+        launcher.beaker.user_name,
+        common_secrets,
+        store_secrets,
+        task_secrets,
+        inject_aws,
+        inject_gcs,
+    )
+
+    job_configs = []
+    experiment_summaries = []
+
+    for exp in experiment_plan:
+        job_config = job_assembler.assemble(exp)
+        job_configs.append(job_config)
+
+        exp_summary = _build_experiment_summary(
+            exp, job_config, task_summary_by_spec, use_async_stream, use_async, launch_config
+        )
+        experiment_summaries.append(exp_summary)
+
+    # Print experiment summaries
+    for exp_summary in experiment_summaries:
         console.print(
             Panel(
-                "\n".join(storage_lines),
-                title="[bold]Storage Configuration[/bold]",
-                border_style="green",
+                Pretty(exp_summary, expand_all=True),
+                title=f"[bold]{exp_summary.name}[/bold]",
+                border_style="cyan",
             )
         )
         console.print()
 
-    # Determine the effective image
-    from olmo_eval.core.constants.infrastructure import BEAKER_DEFAULT_IMAGE
+    # Confirm and launch
+    if not dry_run and not yes and not click.confirm("Proceed with launch?", default=True):
+        console.print("[yellow]Launch cancelled[/yellow]")
+        raise SystemExit(0)
 
-    if image:
-        effective_image = image
-    elif cfg and cfg.beaker_image:
-        effective_image = cfg.beaker_image
-    else:
-        effective_image = BEAKER_DEFAULT_IMAGE
+    launched_experiments: list[str] = []
+    for job_config in job_configs:
+        if not dry_run:
+            experiment = launcher.launch(job_config)
+            if experiment:
+                console.print(f"[green]Launched:[/green] {launcher.experiment_url(experiment)}")
+                launched_experiments.append(experiment.id)
 
-    # Check which groups exist and which need to be created
+    # Follow launched experiments
+    if launched_experiments and not dry_run:
+        _handle_follow(launcher, launched_experiments, follow)
+
+
+def _handle_group_creation(launcher, effective_groups: list[str], dry_run: bool) -> None:
+    """Handle checking and creating Beaker groups."""
     from beaker.exceptions import BeakerGroupNotFound
 
-    existing_groups: list[str] = []
-    missing_groups: list[str] = []
+    existing_groups = []
+    missing_groups = []
 
     for grp in effective_groups:
         qualified_name = f"{launcher.beaker.user_name}/{grp}" if "/" not in grp else grp
@@ -551,160 +441,25 @@ def launch(
                 console.print("[red]Aborted.[/red] Cannot launch without required groups.")
                 raise SystemExit(1) from None
 
+            workspace_obj = launcher.beaker.workspace.get(launcher.workspace)
             for grp in missing_groups:
                 try:
-                    beaker_group = launcher.beaker.group.create(
-                        name=grp,
-                        workspace=workspace_obj,
-                    )
+                    beaker_group = launcher.beaker.group.create(name=grp, workspace=workspace_obj)
                     group_url = launcher.get_group_url(beaker_group)
                     console.print(f"[green]  Created {grp}:[/green] {group_url}")
                 except Exception as e:
                     console.print(f"[red]Error:[/red] Failed to create group '{grp}': {e}")
                     raise SystemExit(1) from None
 
-    # Track launched experiments
-    launched_experiments: list[str] = []
 
-    # Build experiment plan with parallelism and splitting
-    # Group models by compatible runtime to run them in the same job
-    experiment_plan: list[dict] = []
-    split_models: list[str] = []
-
-    def get_model_resources_for_grouping(
-        m_cfg: ModelConfig, m_spec: str
-    ) -> tuple[int, int, str | None, str | None]:
-        """Get runtime-critical resources for a model to determine grouping compatibility."""
-        if cfg is not None:
-            m_resources = cfg.get_model_resources(m_cfg)
-            m_gpus = cli_gpus if cli_gpus is not None else m_resources.get("gpus", 1)
-            m_parallelism = (
-                cli_parallelism
-                if cli_parallelism is not None
-                else m_resources.get("parallelism", 1)
-            )
-            m_cluster = cli_cluster if cli_cluster is not None else m_resources.get("cluster")
-            m_provider = m_resources.get("provider")
-        else:
-            m_gpus = cli_gpus if cli_gpus is not None else (m_cfg.gpus or gpus)
-            m_parallelism = (
-                cli_parallelism
-                if cli_parallelism is not None
-                else (m_cfg.parallelism or parallelism)
-            )
-            m_cluster = cli_cluster if cli_cluster is not None else (m_cfg.cluster or cluster)
-            m_provider = m_cfg.provider
-        return m_gpus, m_parallelism, m_cluster, m_provider
-
-    def get_runtime_signature(m_cfg: ModelConfig, m_spec: str) -> tuple:
-        """Get a hashable runtime signature for grouping compatible models."""
-        m_gpus, m_parallelism, m_cluster, m_provider = get_model_resources_for_grouping(
-            m_cfg, m_spec
-        )
-
-        # Extract inline overrides from spec (e.g., "model::attention_backend=FLASH_ATTN")
-        _, _, inline_overrides = m_spec.partition("::")
-
-        return (m_gpus, m_parallelism, m_cluster, m_provider, inline_overrides)
-
-    # Group models by runtime signature
-    from itertools import groupby
-
-    model_data = [
-        (get_runtime_signature(m_cfg, m_spec), i, m_cfg, m_spec)
-        for i, (m_cfg, m_spec) in enumerate(zip(model_configs, model_specs, strict=True))
-    ]
-    sorted_models = sorted(model_data, key=lambda x: (x[0], x[1]))
-    model_groups = [list(group) for _, group in groupby(sorted_models, key=lambda x: x[0])]
-
-    for model_group in model_groups:
-        group_model_cfgs = [g[2] for g in model_group]
-        group_model_specs = [g[3] for g in model_group]
-
-        # Use first model's resources (all models in group have same signature)
-        first_cfg = group_model_cfgs[0]
-        first_spec = group_model_specs[0]
-        m_gpus, m_parallelism, _, _ = get_model_resources_for_grouping(first_cfg, first_spec)
-
-        # Determine if we have multiple models in this group (for naming)
-        group_has_multiple_models = len(group_model_cfgs) > 1
-
-        for t_priority, t_list in tasks_by_priority.items():
-            # Split tasks into agent and simple categories
-            agent_tasks_in_priority = [t for t in t_list if is_agent_spec(t)]
-            simple_tasks_in_priority = [t for t in t_list if not is_agent_spec(t)]
-
-            # Process each task category (agent tasks get separate experiments)
-            for task_category, category_tasks in [
-                ("agent", agent_tasks_in_priority),
-                ("simple", simple_tasks_in_priority),
-            ]:
-                if not category_tasks:
-                    continue
-
-                # When models are grouped, use base name without model suffix
-                # Only append model suffix if there's exactly one model in the group
-                # and there are multiple model groups (ungrouped models)
-                base_name = name
-                if not group_has_multiple_models and multiple_models:
-                    # Single model in group, but multiple models overall - append model name
-                    short_m = get_model_short_name(first_cfg)
-                    base_name = f"{base_name}-{short_m}"
-                if multiple_priorities:
-                    base_name = f"{base_name}-{t_priority}"
-
-                splits = calculate_experiment_splits(
-                    tasks=category_tasks,
-                    gpus_per_model=m_gpus,
-                    parallelism=m_parallelism,
-                    max_gpus_per_node=max_gpus_per_node,
-                )
-
-                if len(splits) > 1:
-                    for m_cfg in group_model_cfgs:
-                        split_models.append(m_cfg.name_or_path)
-
-                total_splits = len(splits)
-                # Count expanded tasks for this category
-                category_expanded = len(expand_tasks(category_tasks))
-                num_models_in_group = len(group_model_cfgs)
-                for i, split in enumerate(splits):
-                    exp_name = f"{base_name}-{i + 1:03d}" if total_splits > 1 else base_name
-
-                    # Total GPUs = GPUs per split * number of models in group
-                    # Each model needs its own set of GPUs
-                    total_gpus_for_group = split["num_gpus"] * num_models_in_group
-
-                    experiment_plan.append(
-                        {
-                            "name": exp_name,
-                            "model_cfgs": group_model_cfgs,  # List of models in this group
-                            "model_specs": group_model_specs,  # Original specs with ::overrides
-                            "priority": t_priority,
-                            "tasks": split["tasks"],
-                            "original_task_specs": original_task_specs,
-                            "total_expanded_tasks": category_expanded,
-                            "gpus_per_model": m_gpus,
-                            "num_gpus": total_gpus_for_group,
-                            "parallelism": split["parallelism"],
-                            "split_index": i + 1 if total_splits > 1 else None,
-                            "total_splits": total_splits if total_splits > 1 else None,
-                            "is_agent": task_category == "agent",  # Flag for agent tasks
-                        }
-                    )
-
-    # Calculate total expanded tasks
-    total_experiments = len(experiment_plan)
-    total_expanded_tasks = len(valid_tasks) * len(model_configs)
-
-    # Fetch actual task configurations for display
+def _get_task_summaries(valid_tasks: list[str]) -> list[TaskSummary]:
+    """Get task summaries for display."""
     from olmo_eval.evals.tasks import get_task as get_task_instance
     from olmo_eval.evals.tasks.core.registry import parse_task_spec
 
-    task_summaries: list[TaskSummary] = []
+    task_summaries = []
     for task_spec in valid_tasks:
         task_name, variants, inline_overrides = parse_task_spec(task_spec)
-
         try:
             task_instance = get_task_instance(task_spec)
         except ValueError as e:
@@ -719,56 +474,13 @@ def launch(
                 overrides=inline_overrides if inline_overrides else None,
             )
         )
+    return task_summaries
 
-    # Collect required secrets from all task configs
-    all_required_secrets: set[str] = set()
-    for ts in task_summaries:
-        if hasattr(ts.config, "required_secrets") and ts.config.required_secrets:
-            all_required_secrets.update(ts.config.required_secrets)
 
-    # Build model summaries with resolved providers
-    from olmo_eval.core.configs import get_model_config as get_runtime_model_config
-
-    model_summaries: list[ModelSummary] = []
-    for m, m_spec in zip(model_configs, model_specs, strict=True):
-        # Use original spec to extract inline overrides (name_or_path has them stripped)
-        model_base_name, model_inline_overrides = parse_model_spec(m_spec)
-
-        if m.provider:
-            effective_provider = m.provider
-        else:
-            runtime_model_config = get_runtime_model_config(model_base_name)
-            effective_provider = runtime_model_config.provider
-
-        model_summaries.append(
-            ModelSummary(
-                name=model_base_name,
-                gpus=m.gpus or gpus,
-                parallelism=m.parallelism or parallelism,
-                alias=m.alias,
-                provider=effective_provider,
-                overrides=model_inline_overrides if model_inline_overrides else None,
-            )
-        )
-
-    # Build per-experiment summaries for display
-    from olmo_eval.runners import (
-        AgentEvalRunner,
-        AsyncEvalRunner,
-        StreamingEvalRunner,
-        SyncEvalRunner,
-    )
-
-    # Build a lookup for task summaries by spec
-    task_summary_by_spec: dict[str, TaskSummary] = {}
-    for ts in task_summaries:
-        # Build the full spec including variants and overrides
-        full_spec = ts.config.name
-        if ts.spec and ts.spec != ts.config.name:
-            full_spec = ts.spec
-        task_summary_by_spec[full_spec] = ts
-
-    # Ensure common secrets exist in Beaker (needed for BeakerJobConfig)
+def _ensure_secrets(
+    launcher, dry_run: bool, launch_config, all_required_secrets: set[str]
+) -> tuple[list[tuple[str, str]], list[tuple[str, str]], list[tuple[str, str]]]:
+    """Ensure required secrets exist."""
     from olmo_eval.launch.beaker.secrets import (
         ensure_common_secrets,
         ensure_task_secrets,
@@ -777,401 +489,160 @@ def launch(
     )
 
     beaker_username = launcher.beaker.user_name
+
     if dry_run:
-        common_secrets: list[tuple[str, str]] = []
+        common_secrets = []
         if get_local_hf_token():
             common_secrets.append(("HF_TOKEN", f"{beaker_username}_HF_TOKEN"))
         if get_local_wandb_api_key():
             common_secrets.append(("WANDB_API_KEY", f"{beaker_username}_WANDB_API_KEY"))
-        # In dry-run, assume task secrets exist (just show what would be used)
-        task_secrets: list[tuple[str, str]] = [
-            (s, f"{beaker_username}_{s}") for s in sorted(all_required_secrets)
-        ]
+        task_secrets = [(s, f"{beaker_username}_{s}") for s in sorted(all_required_secrets)]
     else:
-        common_secrets = ensure_common_secrets(workspace=workspace)
-        # Validate and collect task-required secrets (will fail if any are missing)
+        common_secrets = ensure_common_secrets(workspace=launch_config.workspace)
         try:
             task_secrets = ensure_task_secrets(
-                workspace=workspace, required_secrets=all_required_secrets
+                workspace=launch_config.workspace, required_secrets=all_required_secrets
             )
         except ValueError as e:
             console.print(f"[red]Error:[/red] {e}")
             raise SystemExit(1) from None
 
-    # Build store secrets list if --store is enabled
-    store_secrets: list[tuple[str, str]] = []
-    if store:
+    store_secrets = []
+    if launch_config.store:
         for env_var in ["PGHOST", "PGPORT", "PGDATABASE", "PGUSER", "PGPASSWORD"]:
             store_secrets.append((env_var, f"olmo_eval_{env_var}"))
 
-    # Print header
-    console.print()
-    console.print(
-        f"[bold]Launching {total_experiments} experiment(s) "
-        f"with {total_expanded_tasks} task(s)[/bold]"
-    )
-    if split_models:
-        console.print(
-            "[dim]  Tasks distributed across multiple experiments due to GPU constraints[/dim]"
-        )
+    return common_secrets, store_secrets, task_secrets
 
-    # Experiment matrix table - show when multiple experiments
-    if total_experiments > 1:
-        matrix_table = Table(show_header=True, title="Experiment Plan")
-        matrix_table.add_column("Name", style="cyan")
-        matrix_table.add_column("Models", style="blue")
-        matrix_table.add_column("Runner", style="magenta")
-        matrix_table.add_column("Priority", style="yellow")
-        matrix_table.add_column("Tasks", justify="right")
-        matrix_table.add_column("GPUs", style="green", justify="right")
 
-        for exp in experiment_plan:
-            task_count = len(exp["tasks"])
-            total_tasks = exp["total_expanded_tasks"]
-            task_display = (
-                f"{task_count}/{total_tasks}" if exp["split_index"] is not None else str(task_count)
-            )
-            model_cfgs = exp["model_cfgs"]
-            if len(model_cfgs) == 1:
-                model_display = model_cfgs[0].name_or_path
-            else:
-                model_display = f"{len(model_cfgs)} models"
-            # Show runner type
-            exp_is_agent = exp.get("is_agent", False)
-            if exp_is_agent:
-                runner_display = "AgentEvalRunner"
-            elif use_async_stream:
-                runner_display = "StreamingEvalRunner"
-            elif use_async:
-                runner_display = "AsyncEvalRunner"
-            else:
-                runner_display = "SyncEvalRunner"
-            matrix_table.add_row(
-                exp["name"],
-                model_display,
-                runner_display,
-                exp["priority"],
-                task_display,
-                str(exp["num_gpus"]),
-            )
-
-        console.print(matrix_table)
-
-    console.print()
-
-    # Build all BeakerJobConfig objects and experiment summaries
-    job_configs: list[BeakerJobConfig] = []
-    experiment_summaries: list[ExperimentSummary] = []
+def _print_experiment_matrix(
+    experiment_plan: list[dict], use_async_stream: bool, use_async: bool
+) -> None:
+    """Print experiment matrix table."""
+    matrix_table = Table(show_header=True, title="Experiment Plan")
+    matrix_table.add_column("Name", style="cyan")
+    matrix_table.add_column("Models", style="blue")
+    matrix_table.add_column("Runner", style="magenta")
+    matrix_table.add_column("Priority", style="yellow")
+    matrix_table.add_column("Tasks", justify="right")
+    matrix_table.add_column("GPUs", style="green", justify="right")
 
     for exp in experiment_plan:
-        exp_model_cfgs = exp["model_cfgs"]
-        exp_model_specs = exp["model_specs"]
-        exp_name = exp["name"]
-        task_list = exp["tasks"]
-        exp_num_gpus = exp["num_gpus"]
-        exp_parallelism = exp["parallelism"]
-        effective_priority = exp["priority"]
+        task_count = len(exp["tasks"])
+        total_tasks = exp["total_expanded_tasks"]
+        task_display = (
+            f"{task_count}/{total_tasks}" if exp["split_index"] is not None else str(task_count)
+        )
+        model_cfgs = exp["model_cfgs"]
+        model_display = (
+            model_cfgs[0].name_or_path if len(model_cfgs) == 1 else f"{len(model_cfgs)} models"
+        )
+
         exp_is_agent = exp.get("is_agent", False)
-
-        # Use first model's resources (all models in group have compatible runtime)
-        first_model_cfg = exp_model_cfgs[0]
-
-        # Get effective resources for this model group
-        if cfg is not None:
-            model_resources = cfg.get_model_resources(first_model_cfg)
-        else:
-            m_para = first_model_cfg.parallelism
-            model_resources = {
-                "gpus": first_model_cfg.gpus if first_model_cfg.gpus is not None else gpus,
-                "parallelism": m_para if m_para is not None else parallelism,
-                "cluster": (
-                    first_model_cfg.cluster if first_model_cfg.cluster is not None else cluster
-                ),
-                "preemptible": (
-                    first_model_cfg.preemptible
-                    if first_model_cfg.preemptible is not None
-                    else preemptible
-                ),
-                "timeout": (
-                    first_model_cfg.timeout if first_model_cfg.timeout is not None else timeout
-                ),
-                "shared_memory": first_model_cfg.shared_memory,
-                "provider": first_model_cfg.provider,
-            }
-
-        # CLI args always override per-model config
-        effective_cluster: str = (
-            cli_cluster if cli_cluster is not None else str(model_resources["cluster"])
-        )
-        effective_preemptible: bool = (
-            cli_preemptible if cli_preemptible is not None else bool(model_resources["preemptible"])
-        )
-        effective_timeout: str = (
-            cli_timeout if cli_timeout is not None else str(model_resources["timeout"])
-        )
-        res_shared_memory = model_resources.get("shared_memory")
-        effective_shared_memory: str = str(res_shared_memory) if res_shared_memory else "10GiB"
-
-        # Build command with all models in this group
-        command: list[str] = ["olmo-eval", "run"]
-
-        # Add each model with its spec (preserves inline overrides like ::attention_backend=X)
-        for m_cfg, m_spec in zip(exp_model_cfgs, exp_model_specs, strict=True):
-            # Check if config file specifies additional inline overrides (load_format, etc.)
-            m_resources = cfg.get_model_resources(m_cfg) if cfg is not None else model_resources
-
-            # Build final model spec with any config-file inline overrides
-            final_model_spec = m_spec
-            config_inline_overrides: list[str] = []
-
-            effective_load_format = m_resources.get("load_format")
-            if effective_load_format:
-                config_inline_overrides.append(f"load_format={effective_load_format}")
-
-            effective_extra_loader_config = m_resources.get("extra_loader_config")
-            if effective_extra_loader_config:
-                json_config = json_module.dumps(
-                    effective_extra_loader_config, separators=(",", ":")
-                )
-                config_inline_overrides.append(f"extra_loader_config={json_config}")
-
-            # Merge config-file overrides with CLI inline overrides
-            if config_inline_overrides:
-                if "::" in final_model_spec:
-                    # Append to existing inline overrides
-                    final_model_spec = f"{final_model_spec},{','.join(config_inline_overrides)}"
-                else:
-                    # Add inline overrides
-                    final_model_spec = f"{final_model_spec}::{','.join(config_inline_overrides)}"
-
-            command.extend(["-m", final_model_spec])
-
-            # Add alias for this model if defined (scoped to the preceding -m flag)
-            if m_cfg.alias:
-                command.extend(["--alias", m_cfg.alias])
-
-        # Add tasks
-        for t in task_list:
-            command.extend(["-t", t])
-
-        # Add parallelism if > 1
-        if exp_parallelism > 1:
-            command.extend(["--parallelism", str(exp_parallelism)])
-
-        # Agent tasks use a dedicated runner with their own vLLM server
         if exp_is_agent:
-            command.append("--agent")
-            # Pass GPU count for tensor parallelism in the agent's vLLM server
-            if exp_num_gpus > 1:
-                command.extend(["--num-gpus", str(exp_num_gpus)])
-        else:
-            # Add async flags if enabled (not applicable to agent tasks)
-            effective_use_async = use_async or model_resources.get("use_async", False)
-            effective_use_async_stream = use_async_stream or model_resources.get(
-                "use_async_stream", False
-            )
-            effective_num_workers = (
-                num_workers if num_workers is not None else model_resources.get("num_workers")
-            )
-            effective_gpus_per_worker = (
-                gpus_per_worker
-                if gpus_per_worker != 1
-                else model_resources.get("gpus_per_worker", 1)
-            )
-
-            if effective_use_async_stream:
-                command.append("--async-stream")
-                if effective_num_workers is not None:
-                    command.extend(["--num-workers", str(effective_num_workers)])
-                if effective_gpus_per_worker and effective_gpus_per_worker != 1:
-                    command.extend(["--gpus-per-worker", str(effective_gpus_per_worker)])
-            elif effective_use_async:
-                command.append("--async")
-                if effective_num_workers is not None:
-                    command.extend(["--num-workers", str(effective_num_workers)])
-                if effective_gpus_per_worker and effective_gpus_per_worker != 1:
-                    command.extend(["--gpus-per-worker", str(effective_gpus_per_worker)])
-
-        # Add S3 options if configured (group is inferred from beaker group)
-        if s3_bucket and s3_prefix:
-            command.extend(["--s3-bucket", s3_bucket])
-            command.extend(["--s3-prefix", s3_prefix])
-            # Use the first beaker group as the S3 group
-            if effective_groups:
-                command.extend(["--s3-group", effective_groups[0]])
-            if s3_endpoint_url:
-                command.extend(["--s3-endpoint-url", s3_endpoint_url])
-            if s3_region != "us-east-1":
-                command.extend(["--s3-region", s3_region])
-
-        # Add experiment group for grouping related experiments
-        if effective_groups:
-            command.extend(["--experiment-group", effective_groups[0]])
-
-        # Add experiment name for database storage
-        command.extend(["--experiment-name", exp_name])
-
-        # Enable PostgreSQL storage backend for persisting evaluation results
-        # (model metrics, task scores, S3 locations) after each task completes.
-        # Credentials are injected via Beaker secrets (PGHOST, PGPORT, etc.)
-        if store:
-            command.append("--store")
-
-        # Debug logging flags
-        if debug_requests:
-            command.append("--debug-requests")
-        if debug_provider:
-            command.append("--debug-provider")
-        if not save_predictions:
-            command.append("--no-save-predictions")
-        if not save_requests:
-            command.append("--no-save-requests")
-
-        # Get the inference provider for this model group (defaults to vllm)
-        # Agent tasks start their own vLLM server, so they need vllm extras
-        from olmo_eval.core.constants.infrastructure import BACKEND_OPTIONAL_GROUPS
-
-        if exp_is_agent:
-            # Agent tasks need vLLM (for server) and agents (for agent SDK)
-            provider_extras = ["vllm", "agents"]
-        else:
-            config_provider = model_resources.get("provider") or "vllm"
-            runtime_provider: str = str(config_provider)
-            provider_group = BACKEND_OPTIONAL_GROUPS.get(runtime_provider)
-            provider_extras = [provider_group] if provider_group else []
-
-        # Combine inference provider and storage dependencies for installation
-        install_extras = list(provider_extras)
-        if store:
-            install_extras.append("postgres")
-
-        # Convert secrets to BeakerEnvSecret objects
-        env_secrets = [
-            BeakerEnvSecret(env_var, secret_name) for env_var, secret_name in common_secrets
-        ]
-        env_secrets.extend(
-            BeakerEnvSecret(env_var, secret_name) for env_var, secret_name in store_secrets
-        )
-        env_secrets.extend(
-            BeakerEnvSecret(env_var, secret_name) for env_var, secret_name in task_secrets
-        )
-
-        # Build env vars: include defaults plus Beaker author for attribution
-        job_env_vars = {
-            "HF_HOME": "/weka/oe-eval-default/oyvindt/hf-cache",
-            "HF_HUB_CACHE": "/weka/oe-eval-default/oyvindt/hf-cache",
-            "BEAKER_AUTHOR": beaker_username,
-        }
-
-        job_config = BeakerJobConfig(
-            name=exp_name,
-            command=command,
-            cluster=effective_cluster,
-            num_gpus=exp_num_gpus,
-            priority=effective_priority,
-            preemptible=effective_preemptible,
-            timeout=effective_timeout,
-            shared_memory=effective_shared_memory,
-            retries=retries,
-            workspace=workspace,
-            budget=budget,
-            extras=install_extras,
-            groups=effective_groups,
-            beaker_image=effective_image,
-            inject_aws_credentials=inject_aws_credentials,
-            inject_gcs_credentials=inject_gcs_credentials,
-            env_vars=job_env_vars,
-            env_secrets=env_secrets,
-        )
-        job_configs.append(job_config)
-
-        # Build experiment summary with the job config
-        # Get model and task summaries for this experiment
-        exp_model_summaries = []
-        for m_cfg, m_spec in zip(exp_model_cfgs, exp_model_specs, strict=True):
-            model_base_name, model_inline_overrides = parse_model_spec(m_spec)
-            exp_model_summaries.append(
-                ModelSummary(
-                    name=model_base_name,
-                    gpus=m_cfg.gpus or gpus,
-                    parallelism=m_cfg.parallelism or parallelism,
-                    alias=m_cfg.alias,
-                    provider=m_cfg.provider,
-                    overrides=model_inline_overrides if model_inline_overrides else None,
-                )
-            )
-
-        exp_task_summaries = []
-        for task_spec in task_list:
-            base_spec = task_spec.rsplit("@", 1)[0] if "@" in task_spec else task_spec
-            if base_spec in task_summary_by_spec:
-                exp_task_summaries.append(task_summary_by_spec[base_spec])
-
-        # Determine runner class for display
-        if exp_is_agent:
-            exp_runner_class = AgentEvalRunner
+            runner_display = "AgentEvalRunner"
         elif use_async_stream:
-            exp_runner_class = StreamingEvalRunner
+            runner_display = "StreamingEvalRunner"
         elif use_async:
-            exp_runner_class = AsyncEvalRunner
+            runner_display = "AsyncEvalRunner"
         else:
-            exp_runner_class = SyncEvalRunner
+            runner_display = "SyncEvalRunner"
 
-        exp_runner_config = RunnerConfig(
-            runner=exp_runner_class,
-            output_dir=BEAKER_RESULT_DIR,
+        matrix_table.add_row(
+            exp["name"],
+            model_display,
+            runner_display,
+            exp["priority"],
+            task_display,
+            str(exp["num_gpus"]),
         )
 
-        experiment_summaries.append(
-            ExperimentSummary(
-                name=exp_name,
-                models=exp_model_summaries,
-                tasks=exp_task_summaries,
-                runner=exp_runner_config,
-                beaker=job_config,
+    console.print(matrix_table)
+
+
+def _build_experiment_summary(
+    exp: dict,
+    job_config,
+    task_summary_by_spec: dict,
+    use_async_stream: bool,
+    use_async: bool,
+    launch_config,
+) -> ExperimentSummary:
+    """Build experiment summary for display."""
+    from olmo_eval.runners import (
+        AgentEvalRunner,
+        AsyncEvalRunner,
+        StreamingEvalRunner,
+        SyncEvalRunner,
+    )
+
+    exp_model_cfgs = exp["model_cfgs"]
+    exp_model_specs = exp["model_specs"]
+    task_list = exp["tasks"]
+    exp_is_agent = exp.get("is_agent", False)
+
+    # Build model summaries
+    exp_model_summaries = []
+    for m_cfg, m_spec in zip(exp_model_cfgs, exp_model_specs, strict=True):
+        model_base_name, model_inline_overrides = parse_model_spec(m_spec)
+        exp_model_summaries.append(
+            ModelSummary(
+                name=model_base_name,
+                gpus=m_cfg.gpus or launch_config.gpus,
+                parallelism=m_cfg.parallelism or launch_config.parallelism,
+                alias=m_cfg.alias,
+                provider=m_cfg.provider,
+                overrides=model_inline_overrides if model_inline_overrides else None,
             )
         )
 
-    # Print experiment summaries with full BeakerJobConfig details
-    for exp_summary in experiment_summaries:
-        console.print(
-            Panel(
-                Pretty(exp_summary, expand_all=True),
-                title=f"[bold]{exp_summary.name}[/bold]",
-                border_style="cyan",
+    # Build task summaries
+    exp_task_summaries = []
+    for task_spec in task_list:
+        base_spec = task_spec.rsplit("@", 1)[0] if "@" in task_spec else task_spec
+        if base_spec in task_summary_by_spec:
+            exp_task_summaries.append(task_summary_by_spec[base_spec])
+
+    # Determine runner class
+    if exp_is_agent:
+        exp_runner_class = AgentEvalRunner
+    elif use_async_stream:
+        exp_runner_class = StreamingEvalRunner
+    elif use_async:
+        exp_runner_class = AsyncEvalRunner
+    else:
+        exp_runner_class = SyncEvalRunner
+
+    exp_runner_config = RunnerConfig(
+        runner=exp_runner_class,
+        output_dir=BEAKER_RESULT_DIR,
+    )
+
+    return ExperimentSummary(
+        name=exp["name"],
+        models=exp_model_summaries,
+        tasks=exp_task_summaries,
+        runner=exp_runner_config,
+        beaker=job_config,
+    )
+
+
+def _handle_follow(launcher, launched_experiments: list[str], follow: bool) -> None:
+    """Handle following launched experiments."""
+    if len(launched_experiments) > 1:
+        console.print(f"\n[bold]Launched {len(launched_experiments)} experiment(s)[/bold]")
+
+    if follow:
+        if len(launched_experiments) == 1:
+            import sys
+
+            exit_code = launcher.follow_experiment(launched_experiments[0])
+            sys.exit(exit_code)
+        else:
+            console.print(
+                "\n[bold]Multiple experiments launched. "
+                "Use 'olmo-eval beaker watch -e <id>' to follow:[/bold]"
             )
-        )
-        console.print()
-
-    # Confirm before launching
-    if not dry_run and not yes and not click.confirm("Proceed with launch?", default=True):
-        console.print("[yellow]Launch cancelled[/yellow]")
-        raise SystemExit(0)
-
-    # Launch experiments
-    for job_config in job_configs:
-        if not dry_run:
-            experiment = launcher.launch(job_config)
-            if experiment:
-                console.print(f"[green]Launched:[/green] {launcher.experiment_url(experiment)}")
-                launched_experiments.append(experiment.id)
-
-    # Summary and follow logic for launched experiments
-    if launched_experiments and not dry_run:
-        if len(launched_experiments) > 1:
-            console.print(f"\n[bold]Launched {len(launched_experiments)} experiment(s)[/bold]")
-
-        if follow:
-            if len(launched_experiments) == 1:
-                import sys
-
-                exit_code = launcher.follow_experiment(launched_experiments[0])
-                sys.exit(exit_code)
-            else:
-                console.print(
-                    "\n[bold]Multiple experiments launched. "
-                    "Use 'olmo-eval beaker watch -e <id>' to follow:[/bold]"
-                )
-                for exp_id in launched_experiments:
-                    url = launcher.get_experiment_url(exp_id)
-                    console.print(f"  - {url}")
+            for exp_id in launched_experiments:
+                url = launcher.get_experiment_url(exp_id)
+                console.print(f"  - {url}")
